[refactor] findAllColorOfMonth() test
