+++ src/test/java/com/emotionmaster/emolog/diary/controller/DiaryControllerTest.java
@@ -127,36 +127,4 @@ class DiaryControllerTest {
         assertThat(diaryList.size()).isZero();
 
     }
-
-    @DisplayName("findAllColorOfMonth() : month 값에 맞춰 월간 색을 모두 조회한다")
-    @Test
-    public void findAllColorOfMonth() throws Exception {
-        final String url = "/api/diary";
-        final String content1 = "content1";
-        final String content7 = "content7";
-        final String color1 = "FFFFFF";
-        final String color7 = "000000";
-        final LocalDate date1 = LocalDate.of(2024,1,20);
-        final LocalDate date7 = LocalDate.now();
-
-        Diary diaryOf7 = diaryRepository.save(new Diary(date7, content7));
-        colorRepository.save(new Color(0, 0, 0, color7, diaryOf7));
-
-        Diary diaryOf1 = diaryRepository.save(new Diary(date1, content1));
-        colorRepository.save(new Color(0, 0, 0, color1, diaryOf1));
-
-        // Test for July (month 7)
-        ResultActions resultsOf7 = mockMvc.perform(get(url)
-                        .param("month", String.valueOf(7)))
-                .andExpect(status().isOk());
-
-        resultsOf7.andExpect(jsonPath("$[0].['" + date7 + "']").value(color7));
-
-        // Test for January (month 1)
-        ResultActions resultsOf1 = mockMvc.perform(get(url)
-                        .param("month", String.valueOf(1)))
-                .andExpect(status().isOk());
-
-        resultsOf1.andExpect(jsonPath("$[0].['" + date1 + "']").value(color1));
-    }
 }
